Restructure strip_chars function
Reworked strip_chars() to return a list of valid file names that can
be used in collect_cocs() to find the correct chain of custodies.
The return value of the function is now a tuple, including a list
of valid file names and a list of bad file names, or None if all
name checks passed.