Add total_file_size fn; rewrite collect cocs
Rewrote collect_cocs (and still rewriting) to enable it to make use of a
list of CoCs that is only constructed once. The previous implementation
would have had the program constructing it once for every CoC needed.
Even if this was, say 5 times, that's too many when it should be able
to be passed once.

Also write a total_file_size function that takes all files in a given
directory any adds their file sizes together. Will be used for
compression statistics at the end of each run.

Began writing recursive function 'aggregator' to handle errors in
searching and collecting CoCs, backchecking PDF ranges and storing
all the data in a dictionary to be added to a master report dict for
eventual processing by ghostscript.

No tests written yet.