Refactor code to clarify and simplify
Add more explicit docstrings for system_checks(), name_check() and
find_coc() that more accurately describe what is being checked for.

Clarify comment on return statement in file_check().

Refactor if/else return statements in name_check to make more pythonic.

Rewrite find_coc (still in progress) to take in a `coc_tuple` consisting
of the same sets that had been used. This is just for simplicity's sake.
It really just adds another layer of abstraction, but I don't like all
the arguments find_coc() takes. This reduces. May be rewritten again.

Add interactive loop for user to allow the program to continue with
report collation if there are missing cocs. Need to make sure that
the pdfs the coc would apply to are taken out of the pdf_stack for
the aggregator function.