Make minor tweaks to collate function
collate() now adds arguments to the command for the subprocess call
correctly. Used to be done with " ".join(list_of_files), but this
creates one large string for ghostscript to take as input; not a list
of separate files to add together.

Also added error handling for missing pdfs (from the back check) to
main() and some error handling for copying reports to BILLINGS.