Fix bug where PDFs were unordered in final report
Bad sorting occurred after On^2 operation in lines 449-452. While
generating a list of matching pdfs, the order gets messed up. This
section of the code really needs to be rewritten, but in the interim, a
simple `list.sort()` is performed before storing the list of matched
pdfs in the aggregated dictionary for the report.